Clean up files deleted 2d stuff solved most warnings

+++ Framework/Base/Source/Enemy/AnimatedEnemy/AnimatedEnemy.cpp
@@ -150,26 +150,26 @@ void CAnimatedEnemy::UpdatesRotationValue(double dt)
 	*/
 	if (bArmRotationPositive)
 	{
-		leftArmRotation += rotationSpeed*dt;
-		rightArmRotation -= rotationSpeed*dt;
+		leftArmRotation += rotationSpeed * static_cast<float>(dt);
+		rightArmRotation -= rotationSpeed * static_cast<float>(dt);
 	}
 	else
 	{
-		leftArmRotation -= rotationSpeed*dt;
-		rightArmRotation += rotationSpeed*dt;
+		leftArmRotation -= rotationSpeed * static_cast<float>(dt);
+		rightArmRotation += rotationSpeed * static_cast<float>(dt);
 	}
 	/*
 	Left leg rotates the opposite direction as right leg
 	*/
 	if (bLegRotationPositive)
 	{
-		leftLegRotation += rotationSpeed*dt;
-		rightLegRotation -= rotationSpeed*dt;
+		leftLegRotation += rotationSpeed * static_cast<float>(dt);
+		rightLegRotation -= rotationSpeed * static_cast<float>(dt);
 	}
 	else
 	{
-		leftLegRotation -= rotationSpeed*dt;
-		rightLegRotation += rotationSpeed*dt;
+		leftLegRotation -= rotationSpeed * static_cast<float>(dt);
+		rightLegRotation += rotationSpeed * static_cast<float>(dt);
 	}
 }
 /*
@@ -225,7 +225,7 @@ void CAnimatedEnemy::Update(double dt)
 					previousPosition = position;
 
 				Vector3 returnToDefaultPosition(defaultPosition - position);
-				position += returnToDefaultPosition.Normalized() * (float)dt * GetAttribute(CAttributes::ATTRIBUTE_TYPES::TYPE_SPEED);
+				position += returnToDefaultPosition.Normalized() * (float)dt * static_cast<float>(GetAttribute(CAttributes::ATTRIBUTE_TYPES::TYPE_SPEED));
 			}
 		}
 	}
@@ -255,7 +255,7 @@ void CAnimatedEnemy::Update(double dt)
 				/*Using comparison of magnitude to mimic the real world environment where if the a person just left you not long ago, you will be more alerted and prepare if the person will return.*/
 				if (displacement.LengthSquared() > 100.f)
 				{
-					position += displacement.Normalized() * (float)dt * GetAttribute(CAttributes::ATTRIBUTE_TYPES::TYPE_SPEED);
+					position += displacement.Normalized() * (float)dt * static_cast<float>(GetAttribute(CAttributes::ATTRIBUTE_TYPES::TYPE_SPEED));
 					previousPosition = position;
 				}
 				else
@@ -300,7 +300,7 @@ void CAnimatedEnemy::Update(double dt)
 				/*Using comparison of magnitude to mimic the real world environment where if the a person just left you not long ago, you will be more alerted and prepare if the person will return.*/
 				if (displacement.LengthSquared() > 50.f)
 				{
-					position += displacement.Normalized() * (float)dt * GetAttribute(CAttributes::ATTRIBUTE_TYPES::TYPE_SPEED);
+					position += displacement.Normalized() * (float)dt * static_cast<float>(GetAttribute(CAttributes::ATTRIBUTE_TYPES::TYPE_SPEED));
 					previousPosition = position;
 				}
 				else
@@ -363,7 +363,7 @@ void CAnimatedEnemy::Update(double dt)
 		}
 
 		if ((nearestPosition - Vector3(position.x, -10.f, position.z)).LengthSquared() > 0.1f)
-			position += directionToGo * (float)dt * GetAttribute(CAttributes::ATTRIBUTE_TYPES::TYPE_SPEED);
+			position += directionToGo * (float)dt * static_cast<float>(GetAttribute(CAttributes::ATTRIBUTE_TYPES::TYPE_SPEED));
 		else
 		{
 			while (path.size() > 0)
